Partitioning pybind11 bindings
