+++ python/src/bindings/util.cpp
@@ -1,0 +1,49 @@
+//  Copyright (c) 2017 Hartmut Kaiser
+//  Copyright (c) 2018 R. Tohid
+//  Copyright (c) 2018 Steven R. Brandt
+//
+//  Distributed under the Boost Software License, Version 1.0. (See accompanying
+//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+#include <phylanx/include/ast.hpp>
+#include <phylanx/include/ir.hpp>
+#include <phylanx/include/util.hpp>
+
+#include <bindings/binding_helpers.hpp>
+#include <bindings/type_casters.hpp>
+
+#include <pybind11/pybind11.h>
+
+///////////////////////////////////////////////////////////////////////////////
+// expose util submodule
+void phylanx::bindings::bind_util(pybind11::module m)
+{
+    auto util = m.def_submodule("util");
+
+    util.def("serialize", &phylanx::bindings::serialize<phylanx::ast::nil>,
+        "serialize an AST optoken object into a byte-stream");
+    util.def("serialize", &phylanx::bindings::serialize<phylanx::ast::optoken>,
+        "serialize an AST optoken object into a byte-stream");
+    util.def("serialize", &phylanx::bindings::serialize<phylanx::ast::identifier>,
+        "serialize an AST identifier object into a byte-stream");
+    util.def("serialize", &phylanx::bindings::serialize<phylanx::ast::primary_expr>,
+        "serialize an AST primary_expr object into a byte-stream");
+    util.def("serialize", &phylanx::bindings::serialize<phylanx::ast::operand>,
+        "serialize an AST operand object into a byte-stream");
+    util.def("serialize", &phylanx::bindings::serialize<phylanx::ast::unary_expr>,
+        "serialize an AST unary_expr object into a byte-stream");
+    util.def("serialize", &phylanx::bindings::serialize<phylanx::ast::operation>,
+        "serialize an AST operation object into a byte-stream");
+    util.def("serialize", &phylanx::bindings::serialize<phylanx::ast::expression>,
+        "serialize an AST expression object into a byte-stream");
+    util.def("serialize", &phylanx::bindings::serialize<phylanx::ast::function_call>,
+        "serialize an AST expression object into a byte-stream");
+
+    util.def("serialize", &phylanx::bindings::serialize<phylanx::ir::node_data<double>>,
+        "serialize a node_data<double> expression object into a byte-stream");
+    util.def("serialize", &phylanx::bindings::serialize<phylanx::ir::node_data<bool>>,
+        "serialize a node_data<bool> expression object into a byte-stream");
+
+    util.def("unserialize", &phylanx::util::unserialize,
+        "un-serialize a byte-stream into a Phylanx object");
+}
